refactor(plugins): move resolve/download/load into the engine + flatten config
The engine now owns turning the `plugins:` config into a live registry: a new
`engine::plugin_load` module holds manifest resolution, the `~/.heph/plugins`
download cache, tilde expansion, and the stabby cdylib loader, exposed as
`Engine::register_plugins(&[PluginSpec])` and `engine::clear_plugin_cache()`.
The bin just registers the in-process factories and hands over the plugin list;
plugin-stabby (host), plugin-abi (convert) and reqwest move from the bin to the
engine crate.

Also drops the `identifier:` wrapper from the config: the plugin source
(builtin/path/url) is now an inline key at the top of each entry alongside
`options:`, matching how it reads. `identifier` is an internal representation
only. PluginSpec gets a hand-rolled Deserialize enforcing exactly-one source and
rejecting unknown fields.
